+++ gremlin/ui/backend.py
@@ -28,15 +28,13 @@ from PySide6.QtCore import Property, Signal, Slot
 import dill
 
 from gremlin import code_runner, common, config, device_initialization, error, \
-    event_handler, mode_manager, profile, shared_state, types
+    event_handler, mode_manager, profile, shared_state
 from gremlin.logical_device import LogicalDevice
 from gremlin.signal import signal
 
-from gremlin.ui.auto_mapper import AutoMapper
 from gremlin.ui.device import InputIdentifier, LogicalDeviceManagementModel
 from gremlin.ui.profile import InputItemModel, ModeHierarchyModel
 from gremlin.ui.script import ScriptListModel
-from gremlin.ui.swap_devices import SwapDevices
 from gremlin.audio_player import AudioPlayer
 
 
@@ -331,14 +329,6 @@ class Backend(QtCore.QObject):
     @Slot(result=LogicalDeviceManagementModel)
     def getLogicalDeviceManagementModel(self) -> LogicalDeviceManagementModel:
         return LogicalDeviceManagementModel(self)
-    
-    @Slot(result=AutoMapper)
-    def getAutoMapper(self) -> AutoMapper:
-        return AutoMapper(self)
-
-    @Slot(result=SwapDevices)
-    def getSwapDevices(self) -> SwapDevices:
-        return SwapDevices(self)
 
     @Slot(str, int, result=bool)
     def isActionExpanded(self, uuid_str: str, index: int) -> bool:
