+++ joystick_gremlin.py
@@ -69,6 +69,7 @@ import gremlin.signal
 
 import gremlin.ui.backend
 import gremlin.ui.option
+import gremlin.ui.tools
 import gremlin.ui.util
